# Use int64 instead of time.Time in timestamp encoder/decoders