###### Required for all non-trivial PRs
- [x] Rebased/mergable
- [x] Tests pass
- [ ] CHANGELOG.md updated

This converts the timestamp encoder to use `int64` instead of `time.Time` to avoid a few unnecessary allocations when encoding and decoding.

```
benchmark                       old ns/op     new ns/op     delta
BenchmarkTimeEncoder            25896         25420         -1.84%
BenchmarkTimeDecoder_Packed     25791         16394         -36.44%
BenchmarkTimeDecoder_RLE        23782         7667          -67.76%

benchmark                       old allocs     new allocs     delta
BenchmarkTimeEncoder            1              1              +0.00%
BenchmarkTimeDecoder_Packed     0              0              +0.00%
BenchmarkTimeDecoder_RLE        1              0              -100.00%

benchmark                       old bytes     new bytes     delta
BenchmarkTimeEncoder            8447          8447          +0.00%
BenchmarkTimeDecoder_Packed     0             0             +0.00%
BenchmarkTimeDecoder_RLE        8192          0             -100.00%
```